Move some things from other programs, add some concerning generating functions
(Herein, for the time being, I am describing only polynomials with factors of the form (p+q*x**n), and thus roots in the form (p/q)**(1/n)*e**(2*π*k/n))
It is known that a linear-recurrent sequence given by a rational generating function (which is like an __iter__ method in terms of a mutating internal structure (the numerator)) has an explicit form for its __getitem__, a(n) in terms of nth exponents of the g.f.'s denominator's roots' reciprocals (which, being an analytic function, is also a 'canonical' way to extend it to nonintegers). The powers of roots each have a coefficient, which is determined by the numerator, though repeated ones' coefficients are not constants but polynomial functions of n, with degree r-1 (where r is the number of repetitions).
Being that the problems of finding n coefficients within a g.f.'s numerator, for which a finite unravelling of the series expansion produces the correct first n terms, and finding the coefficients in the explicit form from this, are systems of n linear equations in n variables, they can be solved in terms of matrix inversion. I am very unknowledgeable regarding this, but as far as I know, this is the most efficient way (and polynomial time isn't bad, I suppose).
Note that in n-dimensional space, the space spanned by the n vectors, where the kth vector's ith component is given by cos(2*π*(i+k)/n), has maximum rank (being that they're equivalent to cyclic permutations of each other (forming the edges of a unit parallelepiped connecting to the origin), so if the average of the vectors is subtracted from each of them, they become the points of the n-dimensional simplex). Due to partial fraction decompositions, multiple sets of roots together (like those based on the square- and cube-roots of unity) remain of maximal rank, providing a number of degrees of freedom equivalent to that of constraints.
However, this same argument applies to the set of vectors with components given by the function (i+k)%n, which allows them to be computed in terms of integer arithmetic.
The next leap of mathematical inelegance (but computational niceness) is that for the nth term, instead of exponentiating these roots of the form (p/q)**(1/deg(f)) (for each factor f) to the power of n, you take (p/q)**(n//deg(f)), then the perturbations of period deg(f) must account for the continuous function's process of growing also.
I haven't done the multiplicities yet, I am committing it because I got it working (as far as I can tell).
Included in the __name__=='__main__' are:
An example of use of the g.f.-solving function for A005418(n+1)=A361870(1,n), the number of length-n binary tapes under symmetry, a rare Pólya enumeration function that is linear-recurrent, for which the non-analytic mode finds a form equivalent to
(2**n+2**⌈n/2⌉)/2,
instead of the g.f.-canonical analytic continuation,
(2**n+(2**(n/2)*(1+(-1)**n)+2**((n+1)/2)*(1-(-1)**n))/2)/2),
And an example of the other kind of matrix inversion for finding the (quite large) g.f. for A055397 (a(n) is the maximum number of cells in an n*n still life), which has a degree-114 numerator and degree-56 denominator. (It takes a few minutes on my computer, maybe you are in the future and it will not for you.)